Changed some files names, minor changes in compressor, major changes in decompressor: it was divided on funcs

+++ matrix_task/src/compressor.cpp
@@ -2,12 +2,18 @@
 #include "element.h"
 #include <vector>
 
-std::vector<struct Element> compress(const std::vector<std::vector<double>> & matrix)
+std::vector<struct Element>
+compress(const std::vector<std::vector<double>>& matrix)
 {
     std::vector<struct Element> result{};
-    for(int i = 0; i < matrix.size(); ++i) {
-        for(int j = 0; j < matrix[i].size(); ++j) {
-            if(matrix[i][j] == 0) continue;
+
+    for(size_t i = 0; i < matrix.size(); ++i)
+    {
+        for(size_t j = 0; j < matrix[i].size(); ++j)
+        {
+            if(matrix[i][j] == 0)
+                continue;
+
             result.push_back(Element(i, j, matrix[i][j]));
         }
     }
